.github/requirements.txt
pandas
PyGithub
copr